Move comment handles to shared parent class
